+++ acclimatise/converter/wdl.py
@@ -25,7 +25,7 @@ def escape_wdl_str(text: str):
     """
     Escape literal quotes in a Python string, to become suitable for WDL
     """
-    return text.replace('"', '\\"')
+    return text.replace('"', '\\"').replace("\n", "\\n")
 
 
 def flag_to_command_input(
@@ -38,7 +38,11 @@ def flag_to_command_input(
         if isinstance(named_flag.arg.args, model.EmptyFlagArg):
             args.update(dict(true=named_flag.arg.longest_synonym, false=""))
         else:
-            args.update(dict(prefix=named_flag.arg.longest_synonym,))
+            args.update(
+                dict(
+                    prefix=named_flag.arg.longest_synonym,
+                )
+            )
     elif isinstance(named_flag, model.Positional):
         args.update(dict(optional=False, position=named_flag.position))